image_sender: check that fs is read-only

+++ image_sender.sh
@@ -26,6 +26,13 @@ function image_send()
     local fs_size=$(get_fs_size_in_kb $fs)
     local fs_size_in_bs=$(( fs_size / bs_k ))
 
+    if ! read_only $fs; then
+        echo "$fs is not remounted as read-only" >&2
+        echo "Run next command to change this:" >&2
+        echo "mount -o remount,ro $fs" >&2
+        return
+    fi
+
     local sender_cmd="dd if=$fs bs=$bs iflag=direct count=$(( fs_size_in_bs + reserve )) | tee >(md5sum >&2)"
     for dst in $(echo $dsts); do
         sender_cmd+=" | tee >(nc -q1 $dst $port)"
